ET_package for Miscanthus data
I started to arrange the "data" and "constants" in "Evapotranspiration" package with our Miscanthus data. Here is the script, please follow the codes and try to continue to finish the arrangement. So we can try to run "Evapotranspiration" functions for our dataset.
P.S. you may find the "DailyET_Miscanthus_2009.2011.csv" file in the Box.