Evap_package use for SABR miscanthus dataset
I arranged the SABR miscanthus data set into the appropriate format to use in the Evapotranspiration package. The file named "DailyET_SABR_miscanthus" is in our class Box folder. Here are the "Evapotranspiration" package codes for the SABR_miscanthus dataset. You can use that code to run the package for the previous miscanthus data set (Rojda_ET_Miscanthus.R) if you have not applied it yet.